# src/main/resources/META-INF/openapi.yaml
openapi: "3.0.3"
info:
  title: "Pokemong API"
  description: "Pokemong API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
paths:
...
  /pokemong/nickname/{nickname}:

    get:
      summary: Get Pokemongs by nickname
      parameters:
        - name: nickname
          in: path
          required: true
          description: The nickname of the Pokemong. It can be a partial nickname. The match is case-insensitive and ignores leading and trailing spaces.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pokemong'
  /pokemong/dob/{startDate}/{endDate}:

    get:
      summary: Get Pokemongs born within a certain date interval
      parameters:
        - name: startDate
          in: path
          required: true
          description: The start of the date interval (inclusive)
          schema:
            type: string
            format: date
        - name: endDate
          in: path
          required: true
          description: The end of the date interval (inclusive)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pokemong'
        '400':
          description: Invalid date format
  /pokemong:

    get:
      summary: Get all Pokemongs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pokemong'

    post:
      summary: Create a new Pokemong
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pokemong'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemong'
        '400':
          description: Non-valid entity
...
components:
  schemas:
    Pokemong:
      type: object
      required:
        - dob
        - level
        - pokedexId
        - evoStage
        - evoTrack
        - types
        - moveSet
        - schemaVersion
      properties:
        nickname:
          type: string
          nullable: true
        dob:
          type: string
          format: date
        level:
          type: integer
          minimum: 1
        pokedexId:
          type: integer
          minimum: 1
        evoStage:
          type: integer
          minimum: 0
        evoTrack:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PokemongName'
        trainer:
          type: string
          description: MongoDB ObjectId referencing a document in a collection
          nullable: true
        types:
          type: array
          minItems: 1
          maxItems: 2
          items:
            $ref: '#/components/schemas/Type'
        moveSet:
          type: array
          maxItems: 4
          items:
            $ref: '#/components/schemas/PokemongMove'
        schemaVersion:
          $ref: '#/components/schemas/SchemaVersion'
...
    Type:
      type: object
      required:
        - name
        - weakAgainst
        - effectiveAgainst
      properties:
        name:
          $ref: '#/components/schemas/TypeName'
        weakAgainst:
          type: array
          items:
            $ref: '#/components/schemas/TypeName'
        effectiveAgainst:
          type: array
          items:
            $ref: '#/components/schemas/TypeName'
    PokemongMove:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
...
    SchemaVersion:
      type: integer
      minimum: 1
      description: must be >= 1, and furthermore must be == latest schema version
...
    TypeName:
      type: string
      enum: [
        NORMAL,
        GRASS,
        ELECTRIC,
        WATER,
        FIRE,
        BUG,
        GHOST,
        PSYCHIC,
        STEEL,
        DARK,
        FLYING,
        ICE,
        POISON,
        GROUND,
        ROCK,
        DRAGON,
        FIGHTING,
        FAIRY
      ]
    PokemongName:
      type: string
      enum: [
        BULBASAUR,
        IVYSAUR,
        VENUSAUR,
        CHARMANDER,
        CHARMELEON,
        CHARIZARD,
        SQUIRTLE,
        WARTORTLE,
        BLASTOISE,
        CATERPIE,
        METAPOD,
        BUTTERFREE,
        WEEDLE,
        KAKUNA,
        BEEDRILL,
        PIDGEY,
        PIDGEOTTO,
        PIDGEOT,
        RATTATA,
        RATICATE,
        SPEAROW,
        FEAROW,
        EKANS,
        ARBOK,
        PIKACHU,
        RAICHU,
        SANDSHREW,
        SANDSLASH,
        NIDORAN_F,
        NIDORINA,
        NIDOQUEEN,
        NIDORAN_M,
        NIDORINO,
        NIDOKING,
        CLEFAIRY,
        CLEFABLE,
        VULPIX,
        NINETALES,
        JIGGLYPUFF,
        WIGGLYTUFF,
        ZUBAT,
        GOLBAT,
        ODDISH,
        GLOOM,
        VILEPLUME,
        PARAS,
        PARASECT,
        VENONAT,
        VENOMOTH,
        DIGLETT,
        DUGTRIO,
        MEOWTH,
        PERSIAN,
        PSYDUCK,
        GOLDUCK,
        MANKEY,
        PRIMEAPE,
        GROWLITHE,
        ARCANINE,
        POLIWAG,
        POLIWHIRL,
        POLIWRATH,
        ABRA,
        KADABRA,
        ALAKAZAM,
        MACHOP,
        MACHOKE,
        MACHAMP,
        BELLSPROUT,
        WEEPINBELL,
        VICTREEBEL,
        TENTACOOL,
        TENTACRUEL,
        GEODUDE,
        GRAVELER,
        GOLEM,
        PONYTA,
        RAPIDASH,
        SLOWPOKE,
        SLOWBRO,
        MAGNEMITE,
        MAGNETON,
        FARFETCHD,
        DODUO,
        DODRIO,
        SEEL,
        DEWGONG,
        GRIMER,
        MUK,
        SHELLDER,
        CLOYSTER,
        GASTLY,
        HAUNTER,
        GENGAR,
        ONIX,
        DROWZEE,
        HYPNO,
        KRABBY,
        KINGLER,
        VOLTORB,
        ELECTRODE,
        EXEGGCUTE,
        EXEGGUTOR,
        CUBONE,
        MAROWAK,
        HITMONLEE,
        HITMONCHAN,
        LICKITUNG,
        KOFFING,
        WEEZING,
        RHYHORN,
        RHYDON,
        CHANSEY,
        TANGELA,
        KANGASKHAN,
        HORSEA,
        SEADRA,
        GOLDEEN,
        SEAKING,
        STARYU,
        STARMIE,
        MR_MIME,
        SCYTHER,
        JYNX,
        ELECTABUZZ,
        MAGMAR,
        PINSIR,
        TAUROS,
        MAGIKARP,
        GYARADOS,
        LAPRAS,
        DITTO,
        EEVEE,
        VAPOREON,
        JOLTEON,
        FLAREON,
        PORYGON,
        OMANYTE,
        OMASTAR,
        KABUTO,
        KABUTOPS,
        AERODACTYL,
        SNORLAX,
        ARTICUNO,
        ZAPDOS,
        MOLTRES,
        DRATINI,
        DRAGONAIR,
        DRAGONITE,
        MEWTWO,
        MEW,
        CHIKORITA,
        BAYLEEF,
        MEGANIUM,
        CYNDAQUIL,
        QUILAVA,
        TYPHLOSION,
        TOTODILE,
        CROCONAW,
        FERALIGATR,
        SENTRET,
        FURRET,
        HOOTHOOT,
        NOCTOWL,
        LEDYBA,
        LEDIAN,
        SPINARAK,
        ARIADOS,
        CROBAT,
        CHINCHOU,
        LANTURN,
        PICHU,
        CLEFFA,
        IGGLYBUFF,
        TOGEPI,
        TOGETIC,
        NATU,
        XATU,
        MAREEP,
        FLAAFFY,
        AMPHAROS,
        BELLOSSOM,
        MARILL,
        AZUMARILL,
        SUDOWOODO,
        POLITOED,
        HOPPIP,
        SKIPLOOM,
        JUMPLUFF,
        AIPOM,
        SUNKERN,
        SUNFLORA,
        YANMA,
        WOOPER,
        QUAGSIRE,
        ESPEON,
        UMBREON,
        MURKROW,
        SLOWKING,
        MISDREAVUS,
        UNOWN,
        WOBBUFFET,
        GIRAFARIG,
        PINECO,
        FORRETRESS,
        DUNSPARCE,
        GLIGAR,
        STEELIX,
        SNUBBULL,
        GRANBULL,
        QWILFISH,
        SCIZOR,
        SHUCKLE,
        HERACROSS,
        SNEASEL,
        TEDDIURSA,
        URSARING,
        SLUGMA,
        MAGCARGO,
        SWINUB,
        PILOSWINE,
        CORSOLA,
        REMORAID,
        OCTILLERY,
        DELIBIRD,
        MANTINE,
        SKARMORY,
        HOUNDOUR,
        HOUNDOOM,
        KINGDRA,
        PHANPY,
        DONPHAN,
        PORYGON2,
        STANTLER,
        SMEARGLE,
        TYROGUE,
        HITMONTOP,
        SMOOCHUM,
        ELEKID,
        MAGBY,
        MILTANK,
        BLISSEY,
        RAIKOU,
        ENTEI,
        SUICUNE,
        LARVITAR,
        PUPITAR,
        TYRANITAR,
        LUGIA,
        HO_OH,
        CELEBI,
        TREECKO,
        GROVYLE,
        SCEPTILE,
        TORCHIC,
        COMBUSKEN,
        BLAZIKEN,
        MUDKIP,
        MARSHTOMP,
        SWAMPERT,
        POOCHYENA,
        MIGHTYENA,
        ZIGZAGOON,
        LINOONE,
        WURMPLE,
        SILCOON,
        BEAUTIFLY,
        CASCOON,
        DUSTOX,
        LOTAD,
        LOMBRE,
        LUDICOLO,
        SEEDOT,
        NUZLEAF,
        SHIFTRY,
        TAILLOW,
        SWELLOW,
        WINGULL,
        PELIPPER,
        RALTS,
        KIRLIA,
        GARDEVOIR,
        SURSKIT,
        MASQUERAIN,
        SHROOMISH,
        BRELOOM,
        SLAKOTH,
        VIGOROTH,
        SLAKING,
        NINCADA,
        NINJASK,
        SHEDINJA,
        WHISMUR,
        LOUDRED,
        EXPLOUD,
        MAKUHITA,
        HARIYAMA,
        AZURILL,
        NOSEPASS,
        SKITTY,
        DELCATTY,
        SABLEYE,
        MAWILE,
        ARON,
        LAIRON,
        AGGRON,
        MEDITITE,
        MEDICHAM,
        ELECTRIKE,
        MANECTRIC,
        PLUSLE,
        MINUN,
        VOLBEAT,
        ILLUMISE,
        ROSELIA,
        GULPIN,
        SWALOT,
        CARVANHA,
        SHARPEDO,
        WAILMER,
        WAILORD,
        NUMEL,
        CAMERUPT,
        TORKOAL,
        SPOINK,
        GRUMPIG,
        SPINDA,
        TRAPINCH,
        VIBRAVA,
        FLYGON,
        CACNEA,
        CACTURNE,
        SWABLU,
        ALTARIA,
        ZANGOOSE,
        SEVIPER,
        LUNATONE,
        SOLROCK,
        BARBOACH,
        WHISCASH,
        CORPHISH,
        CRAWDAUNT,
        BALTOY,
        CLAYDOL,
        LILEEP,
        CRADILY,
        ANORITH,
        ARMALDO,
        FEEBAS,
        MILOTIC,
        CASTFORM,
        KECLEON,
        SHUPPET,
        BANETTE,
        DUSKULL,
        DUSCLOPS,
        TROPIUS,
        CHIMECHO,
        ABSOL,
        WYNAUT,
        SNORUNT,
        GLALIE,
        SPHEAL,
        SEALEO,
        WALREIN,
        CLAMPERL,
        HUNTAIL,
        GOREBYSS,
        RELICANTH,
        LUVDISC,
        BAGON,
        SHELGON,
        SALAMENCE,
        BELDUM,
        METANG,
        METAGROSS,
        REGIROCK,
        REGICE,
        REGISTEEL,
        LATIAS,
        LATIOS,
        KYOGRE,
        GROUDON,
        RAYQUAZA,
        JIRACHI,
        DEOXYS_NORMAL,
        TURTWIG,
        GROTLE,
        TORTERRA,
        CHIMCHAR,
        MONFERNO,
        INFERNAPE,
        PIPLUP,
        PRINPLUP,
        EMPOLEON,
        STARLY,
        STARAVIA,
        STARAPTOR,
        BIDOOF,
        BIBAREL,
        KRICKETOT,
        KRICKETUNE,
        SHINX,
        LUXIO,
        LUXRAY,
        BUDEW,
        ROSERADE,
        CRANIDOS,
        RAMPARDOS,
        SHIELDON,
        BASTIODON,
        BURMY,
        WORMADAM_PLANT,
        MOTHIM,
        COMBEE,
        VESPIQUEN,
        PACHIRISU,
        BUIZEL,
        FLOATZEL,
        CHERUBI,
        CHERRIM,
        SHELLOS,
        GASTRODON,
        AMBIPOM,
        DRIFLOON,
        DRIFBLIM,
        BUNEARY,
        LOPUNNY,
        MISMAGIUS,
        HONCHKROW,
        GLAMEOW,
        PURUGLY,
        CHINGLING,
        STUNKY,
        SKUNTANK,
        BRONZOR,
        BRONZONG,
        BONSLY,
        MIME_JR,
        HAPPINY,
        CHATOT,
        SPIRITOMB,
        GIBLE,
        GABITE,
        GARCHOMP,
        MUNCHLAX,
        RIOLU,
        LUCARIO,
        HIPPOPOTAS,
        HIPPOWDON,
        SKORUPI,
        DRAPION,
        CROAGUNK,
        TOXICROAK,
        CARNIVINE,
        FINNEON,
        LUMINEON,
        MANTYKE,
        SNOVER,
        ABOMASNOW,
        WEAVILE,
        MAGNEZONE,
        LICKILICKY,
        RHYPERIOR,
        TANGROWTH,
        ELECTIVIRE,
        MAGMORTAR,
        TOGEKISS,
        YANMEGA,
        LEAFEON,
        GLACEON,
        GLISCOR,
        MAMOSWINE,
        PORYGON_Z,
        GALLADE,
        PROBOPASS,
        DUSKNOIR,
        FROSLASS,
        ROTOM,
        UXIE,
        MESPRIT,
        AZELF,
        DIALGA,
        PALKIA,
        HEATRAN,
        REGIGIGAS,
        GIRATINA_ALTERED,
        CRESSELIA,
        PHIONE,
        MANAPHY,
        DARKRAI,
        SHAYMIN_LAND,
        ARCEUS,
        VICTINI,
        SNIVY,
        SERVINE,
        SERPERIOR,
        TEPIG,
        PIGNITE,
        EMBOAR,
        OSHAWOTT,
        DEWOTT,
        SAMUROTT,
        PATRAT,
        WATCHOG,
        LILLIPUP,
        HERDIER,
        STOUTLAND,
        PURRLOIN,
        LIEPARD,
        PANSAGE,
        SIMISAGE,
        PANSEAR,
        SIMISEAR,
        PANPOUR,
        SIMIPOUR,
        MUNNA,
        MUSHARNA,
        PIDOVE,
        TRANQUILL,
        UNFEZANT,
        BLITZLE,
        ZEBSTRIKA,
        ROGGENROLA,
        BOLDORE,
        GIGALITH,
        WOOBAT,
        SWOOBAT,
        DRILBUR,
        EXCADRILL,
        AUDINO,
        TIMBURR,
        GURDURR,
        CONKELDURR,
        TYMPOLE,
        PALPITOAD,
        SEISMITOAD,
        THROH,
        SAWK,
        SEWADDLE,
        SWADLOON,
        LEAVANNY,
        VENIPEDE,
        WHIRLIPEDE,
        SCOLIPEDE,
        COTTONEE,
        WHIMSICOTT,
        PETILIL,
        LILLIGANT,
        BASCULIN_RED_STRIPED,
        SANDILE,
        KROKOROK,
        KROOKODILE,
        DARUMAKA,
        DARMANITAN_STANDARD,
        MARACTUS,
        DWEBBLE,
        CRUSTLE,
        SCRAGGY,
        SCRAFTY,
        SIGILYPH,
        YAMASK,
        COFAGRIGUS,
        TIRTOUGA,
        CARRACOSTA,
        ARCHEN,
        ARCHEOPS,
        TRUBBISH,
        GARBODOR,
        ZORUA,
        ZOROARK,
        MINCCINO,
        CINCCINO,
        GOTHITA,
        GOTHORITA,
        GOTHITELLE,
        SOLOSIS,
        DUOSION,
        REUNICLUS,
        DUCKLETT,
        SWANNA,
        VANILLITE,
        VANILLISH,
        VANILLUXE,
        DEERLING,
        SAWSBUCK,
        EMOLGA,
        KARRABLAST,
        ESCAVALIER,
        FOONGUS,
        AMOONGUSS,
        FRILLISH,
        JELLICENT,
        ALOMOMOLA,
        JOLTIK,
        GALVANTULA,
        FERROSEED,
        FERROTHORN,
        KLINK,
        KLANG,
        KLINKLANG,
        TYNAMO,
        EELEKTRIK,
        EELEKTROSS,
        ELGYEM,
        BEHEEYEM,
        LITWICK,
        LAMPENT,
        CHANDELURE,
        AXEW,
        FRAXURE,
        HAXORUS,
        CUBCHOO,
        BEARTIC,
        CRYOGONAL,
        SHELMET,
        ACCELGOR,
        STUNFISK,
        MIENFOO,
        MIENSHAO,
        DRUDDIGON,
        GOLETT,
        GOLURK,
        PAWNIARD,
        BISHARP,
        BOUFFALANT,
        RUFFLET,
        BRAVIARY,
        VULLABY,
        MANDIBUZZ,
        HEATMOR,
        DURANT,
        DEINO,
        ZWEILOUS,
        HYDREIGON,
        LARVESTA,
        VOLCARONA,
        COBALION,
        TERRAKION,
        VIRIZION,
        TORNADUS_INCARNATE,
        THUNDURUS_INCARNATE,
        RESHIRAM,
        ZEKROM,
        LANDORUS_INCARNATE,
        KYUREM,
        KELDEO_ORDINARY,
        MELOETTA_ARIA,
        GENESECT,
        CHESPIN,
        QUILLADIN,
        CHESNAUGHT,
        FENNEKIN,
        BRAIXEN,
        DELPHOX,
        FROAKIE,
        FROGADIER,
        GRENINJA,
        BUNNELBY,
        DIGGERSBY,
        FLETCHLING,
        FLETCHINDER,
        TALONFLAME,
        SCATTERBUG,
        SPEWPA,
        VIVILLON,
        LITLEO,
        PYROAR,
        FLABEBE,
        FLOETTE,
        FLORGES,
        SKIDDO,
        GOGOAT,
        PANCHAM,
        PANGORO,
        FURFROU,
        ESPURR,
        MEOWSTIC_MALE,
        HONEDGE,
        DOUBLADE,
        AEGISLASH_SHIELD,
        SPRITZEE,
        AROMATISSE,
        SWIRLIX,
        SLURPUFF,
        INKAY,
        MALAMAR,
        BINACLE,
        BARBARACLE,
        SKRELP,
        DRAGALGE,
        CLAUNCHER,
        CLAWITZER,
        HELIOPTILE,
        HELIOLISK,
        TYRUNT,
        TYRANTRUM,
        AMAURA,
        AURORUS,
        SYLVEON,
        HAWLUCHA,
        DEDENNE,
        CARBINK,
        GOOMY,
        SLIGGOO,
        GOODRA,
        KLEFKI,
        PHANTUMP,
        TREVENANT,
        PUMPKABOO_AVERAGE,
        GOURGEIST_AVERAGE,
        BERGMITE,
        AVALUGG,
        NOIBAT,
        NOIVERN,
        XERNEAS,
        YVELTAL,
        ZYGARDE_50,
        DIANCIE,
        HOOPA,
        VOLCANION,
        ROWLET,
        DARTRIX,
        DECIDUEYE,
        LITTEN,
        TORRACAT,
        INCINEROAR,
        POPPLIO,
        BRIONNE,
        PRIMARINA,
        PIKIPEK,
        TRUMBEAK,
        TOUCANNON,
        YUNGOOS,
        GUMSHOOS,
        GRUBBIN,
        CHARJABUG,
        VIKAVOLT,
        CRABRAWLER,
        CRABOMINABLE,
        ORICORIO_BAILE,
        CUTIEFLY,
        RIBOMBEE,
        ROCKRUFF,
        LYCANROC_MIDDAY,
        WISHIWASHI_SOLO,
        MAREANIE,
        TOXAPEX,
        MUDBRAY,
        MUDSDALE,
        DEWPIDER,
        ARAQUANID,
        FOMANTIS,
        LURANTIS,
        MORELULL,
        SHIINOTIC,
        SALANDIT,
        SALAZZLE,
        STUFFUL,
        BEWEAR,
        BOUNSWEET,
        STEENEE,
        TSAREENA,
        COMFEY,
        ORANGURU,
        PASSIMIAN,
        WIMPOD,
        GOLISOPOD,
        SANDYGAST,
        PALOSSAND,
        PYUKUMUKU,
        TYPE_NULL,
        SILVALLY,
        MINIOR_RED_METEOR,
        KOMALA,
        TURTONATOR,
        TOGEDEMARU,
        MIMIKYU_DISGUISED,
        BRUXISH,
        DRAMPA,
        DHELMISE,
        JANGMO_O,
        HAKAMO_O,
        KOMMO_O,
        TAPU_KOKO,
        TAPU_LELE,
        TAPU_BULU,
        TAPU_FINI,
        COSMOG,
        COSMOEM,
        SOLGALEO,
        LUNALA,
        NIHILEGO,
        BUZZWOLE,
        PHEROMOSA,
        XURKITREE,
        CELESTEELA,
        KARTANA,
        GUZZLORD,
        NECROZMA,
        MAGEARNA,
        MARSHADOW,
        POIPOLE,
        NAGANADEL,
        STAKATAKA,
        BLACEPHALON,
        ZERAORA,
        MELTAN,
        MELMETAL,
        GROOKEY,
        THWACKEY,
        RILLABOOM,
        SCORBUNNY,
        RABOOT,
        CINDERACE,
        SOBBLE,
        DRIZZILE,
        INTELEON,
        SKWOVET,
        GREEDENT,
        ROOKIDEE,
        CORVISQUIRE,
        CORVIKNIGHT,
        BLIPBUG,
        DOTTLER,
        ORBEETLE,
        NICKIT,
        THIEVUL,
        GOSSIFLEUR,
        ELDEGOSS,
        WOOLOO,
        DUBWOOL,
        CHEWTLE,
        DREDNAW,
        YAMPER,
        BOLTUND,
        ROLYCOLY,
        CARKOL,
        COALOSSAL,
        APPLIN,
        FLAPPLE,
        APPLETUN,
        SILICOBRA,
        SANDACONDA,
        CRAMORANT,
        ARROKUDA,
        BARRASKEWDA,
        TOXEL,
        TOXTRICITY_AMPED,
        SIZZLIPEDE,
        CENTISKORCH,
        CLOBBOPUS,
        GRAPPLOCT,
        SINISTEA,
        POLTEAGEIST,
        HATENNA,
        HATTREM,
        HATTERENE,
        IMPIDIMP,
        MORGREM,
        GRIMMSNARL,
        OBSTAGOON,
        PERRSERKER,
        CURSOLA,
        SIRFETCHD,
        MR_RIME,
        RUNERIGUS,
        MILCERY,
        ALCREMIE,
        FALINKS,
        PINCURCHIN,
        SNOM,
        FROSMOTH,
        STONJOURNER,
        EISCUE_ICE,
        INDEEDEE_MALE,
        MORPEKO_FULL_BELLY,
        CUFANT,
        COPPERAJAH,
        DRACOZOLT,
        ARCTOZOLT,
        DRACOVISH,
        ARCTOVISH,
        DURALUDON,
        DREEPY,
        DRAKLOAK,
        DRAGAPULT,
        ZACIAN,
        ZAMAZENTA,
        ETERNATUS,
        KUBFU,
        URSHIFU_SINGLE_STRIKE,
        ZARUDE,
        REGIELEKI,
        REGIDRAGO,
        GLASTRIER,
        SPECTRIER,
        CALYREX,
        WYRDEER,
        KLEAVOR,
        URSALUNA,
        BASCULEGION_MALE,
        SNEASLER,
        OVERQWIL,
        ENAMORUS_INCARNATE,
        SPRIGATITO,
        FLORAGATO,
        MEOWSCARADA,
        FUECOCO,
        CROCALOR,
        SKELEDIRGE,
        QUAXLY,
        QUAXWELL,
        QUAQUAVAL,
        LECHONK,
        OINKOLOGNE,
        TAROUNTULA,
        SPIDOPS,
        NYMBLE,
        LOKIX,
        PAWMI,
        PAWMO,
        PAWMOT,
        TANDEMAUS,
        MAUSHOLD,
        FIDOUGH,
        DACHSBUN,
        SMOLIV,
        DOLLIV,
        ARBOLIVA,
        SQUAWKABILLY,
        NACLI,
        NACLSTACK,
        GARGANACL,
        CHARCADET,
        ARMAROUGE,
        CERULEDGE,
        TADBULB,
        BELLIBOLT,
        WATTREL,
        KILOWATTREL,
        MASCHIFF,
        MABOSSTIFF,
        SHROODLE,
        GRAFAIAI,
        BRAMBLIN,
        BRAMBLEGHAST,
        TOEDSCOOL,
        TOEDSCRUEL,
        KLAWF,
        CAPSAKID,
        SCOVILLAIN,
        RELLOR,
        RABSCA,
        FLITTLE,
        ESPATHRA,
        TINKATINK,
        TINKATUFF,
        TINKATON,
        WIGLETT,
        WUGTRIO,
        BOMBIRDIER,
        FINIZEN,
        PALAFIN,
        VAROOM,
        REVAVROOM,
        CYCLIZAR,
        ORTHWORM,
        GLIMMET,
        GLIMMORA,
        GREAVARD,
        HOUNDSTONE,
        FLAMIGO,
        CETODDLE,
        CETITAN,
        VELUZA,
        DONDOZO,
        TATSUGIRI,
        ANNIHILAPE,
        CLODSIRE,
        FARIGIRAF,
        DUDUNSPARCE,
        KINGAMBIT,
        GREAT_TUSK,
        SCREAM_TAIL,
        BRUTE_BONNET,
        FLUTTER_MANE,
        SLITHER_WING,
        SANDY_SHOCKS,
        IRON_TREADS,
        IRON_BUNDLE,
        IRON_HANDS,
        IRON_JUGULIS,
        IRON_MOTH,
        IRON_THORNS,
        FRIGIBAX,
        ARCTIBAX,
        BAXCALIBUR,
        GIMMIGHOUL,
        GHOLDENGO,
        WO_CHIEN,
        CHIEN_PAO,
        TING_LU,
        CHI_YU,
        ROARING_MOON,
        IRON_VALIANT,
        KORAIDON,
        MIRAIDON,
        WALKING_WAKE,
        IRON_LEAVES,
        DEOXYS_ATTACK,
        DEOXYS_DEFENSE,
        DEOXYS_SPEED,
        WORMADAM_SANDY,
        WORMADAM_TRASH,
        SHAYMIN_SKY,
        GIRATINA_ORIGIN,
        ROTOM_HEAT,
        ROTOM_WASH,
        ROTOM_FROST,
        ROTOM_FAN,
        ROTOM_MOW,
        CASTFORM_SUNNY,
        CASTFORM_RAINY,
        CASTFORM_SNOWY,
        BASCULIN_BLUE_STRIPED,
        DARMANITAN_ZEN,
        MELOETTA_PIROUETTE,
        TORNADUS_THERIAN,
        THUNDURUS_THERIAN,
        LANDORUS_THERIAN,
        KYUREM_BLACK,
        KYUREM_WHITE,
        KELDEO_RESOLUTE,
        MEOWSTIC_FEMALE,
        AEGISLASH_BLADE,
        PUMPKABOO_SMALL,
        PUMPKABOO_LARGE,
        PUMPKABOO_SUPER,
        GOURGEIST_SMALL,
        GOURGEIST_LARGE,
        GOURGEIST_SUPER,
        VENUSAUR_MEGA,
        CHARIZARD_MEGA_X,
        CHARIZARD_MEGA_Y,
        BLASTOISE_MEGA,
        ALAKAZAM_MEGA,
        GENGAR_MEGA,
        KANGASKHAN_MEGA,
        PINSIR_MEGA,
        GYARADOS_MEGA,
        AERODACTYL_MEGA,
        MEWTWO_MEGA_X,
        MEWTWO_MEGA_Y,
        AMPHAROS_MEGA,
        SCIZOR_MEGA,
        HERACROSS_MEGA,
        HOUNDOOM_MEGA,
        TYRANITAR_MEGA,
        BLAZIKEN_MEGA,
        GARDEVOIR_MEGA,
        MAWILE_MEGA,
        AGGRON_MEGA,
        MEDICHAM_MEGA,
        MANECTRIC_MEGA,
        BANETTE_MEGA,
        ABSOL_MEGA,
        GARCHOMP_MEGA,
        LUCARIO_MEGA,
        ABOMASNOW_MEGA,
        FLOETTE_ETERNAL,
        LATIAS_MEGA,
        LATIOS_MEGA,
        SWAMPERT_MEGA,
        SCEPTILE_MEGA,
        SABLEYE_MEGA,
        ALTARIA_MEGA,
        GALLADE_MEGA,
        AUDINO_MEGA,
        SHARPEDO_MEGA,
        SLOWBRO_MEGA,
        STEELIX_MEGA,
        PIDGEOT_MEGA,
        GLALIE_MEGA,
        DIANCIE_MEGA,
        METAGROSS_MEGA,
        KYOGRE_PRIMAL,
        GROUDON_PRIMAL,
        RAYQUAZA_MEGA,
        PIKACHU_ROCK_STAR,
        PIKACHU_BELLE,
        PIKACHU_POP_STAR,
        PIKACHU_PHD,
        PIKACHU_LIBRE,
        PIKACHU_COSPLAY,
        HOOPA_UNBOUND,
        CAMERUPT_MEGA,
        LOPUNNY_MEGA,
        SALAMENCE_MEGA,
        BEEDRILL_MEGA,
        RATTATA_ALOLA,
        RATICATE_ALOLA,
        RATICATE_TOTEM_ALOLA,
        PIKACHU_ORIGINAL_CAP,
        PIKACHU_HOENN_CAP,
        PIKACHU_SINNOH_CAP,
        PIKACHU_UNOVA_CAP,
        PIKACHU_KALOS_CAP,
        PIKACHU_ALOLA_CAP,
        RAICHU_ALOLA,
        SANDSHREW_ALOLA,
        SANDSLASH_ALOLA,
        VULPIX_ALOLA,
        NINETALES_ALOLA,
        DIGLETT_ALOLA,
        DUGTRIO_ALOLA,
        MEOWTH_ALOLA,
        PERSIAN_ALOLA,
        GEODUDE_ALOLA,
        GRAVELER_ALOLA,
        GOLEM_ALOLA,
        GRIMER_ALOLA,
        MUK_ALOLA,
        EXEGGUTOR_ALOLA,
        MAROWAK_ALOLA,
        GRENINJA_BATTLE_BOND,
        GRENINJA_ASH,
        ZYGARDE_10_POWER_CONSTRUCT,
        ZYGARDE_50_POWER_CONSTRUCT,
        ZYGARDE_COMPLETE,
        GUMSHOOS_TOTEM,
        VIKAVOLT_TOTEM,
        ORICORIO_POM_POM,
        ORICORIO_PAU,
        ORICORIO_SENSU,
        LYCANROC_MIDNIGHT,
        WISHIWASHI_SCHOOL,
        LURANTIS_TOTEM,
        SALAZZLE_TOTEM,
        MINIOR_ORANGE_METEOR,
        MINIOR_YELLOW_METEOR,
        MINIOR_GREEN_METEOR,
        MINIOR_BLUE_METEOR,
        MINIOR_INDIGO_METEOR,
        MINIOR_VIOLET_METEOR,
        MINIOR_RED,
        MINIOR_ORANGE,
        MINIOR_YELLOW,
        MINIOR_GREEN,
        MINIOR_BLUE,
        MINIOR_INDIGO,
        MINIOR_VIOLET,
        MIMIKYU_BUSTED,
        MIMIKYU_TOTEM_DISGUISED,
        MIMIKYU_TOTEM_BUSTED,
        KOMMO_O_TOTEM,
        MAGEARNA_ORIGINAL,
        PIKACHU_PARTNER_CAP,
        MAROWAK_TOTEM,
        RIBOMBEE_TOTEM,
        ROCKRUFF_OWN_TEMPO,
        LYCANROC_DUSK,
        ARAQUANID_TOTEM,
        TOGEDEMARU_TOTEM,
        NECROZMA_DUSK,
        NECROZMA_DAWN,
        NECROZMA_ULTRA,
        PIKACHU_STARTER,
        EEVEE_STARTER,
        PIKACHU_WORLD_CAP,
        MEOWTH_GALAR,
        PONYTA_GALAR,
        RAPIDASH_GALAR,
        SLOWPOKE_GALAR,
        SLOWBRO_GALAR,
        FARFETCHD_GALAR,
        WEEZING_GALAR,
        MR_MIME_GALAR,
        ARTICUNO_GALAR,
        ZAPDOS_GALAR,
        MOLTRES_GALAR,
        SLOWKING_GALAR,
        CORSOLA_GALAR,
        ZIGZAGOON_GALAR,
        LINOONE_GALAR,
        DARUMAKA_GALAR,
        DARMANITAN_GALAR_STANDARD,
        DARMANITAN_GALAR_ZEN,
        YAMASK_GALAR,
        STUNFISK_GALAR,
        ZYGARDE_10,
        CRAMORANT_GULPING,
        CRAMORANT_GORGING,
        TOXTRICITY_LOW_KEY,
        EISCUE_NOICE,
        INDEEDEE_FEMALE,
        MORPEKO_HANGRY,
        ZACIAN_CROWNED,
        ZAMAZENTA_CROWNED,
        ETERNATUS_ETERNAMAX,
        URSHIFU_RAPID_STRIKE,
        ZARUDE_DADA,
        CALYREX_ICE,
        CALYREX_SHADOW,
        VENUSAUR_GMAX,
        CHARIZARD_GMAX,
        BLASTOISE_GMAX,
        BUTTERFREE_GMAX,
        PIKACHU_GMAX,
        MEOWTH_GMAX,
        MACHAMP_GMAX,
        GENGAR_GMAX,
        KINGLER_GMAX,
        LAPRAS_GMAX,
        EEVEE_GMAX,
        SNORLAX_GMAX,
        GARBODOR_GMAX,
        MELMETAL_GMAX,
        RILLABOOM_GMAX,
        CINDERACE_GMAX,
        INTELEON_GMAX,
        CORVIKNIGHT_GMAX,
        ORBEETLE_GMAX,
        DREDNAW_GMAX,
        COALOSSAL_GMAX,
        FLAPPLE_GMAX,
        APPLETUN_GMAX,
        SANDACONDA_GMAX,
        TOXTRICITY_AMPED_GMAX,
        CENTISKORCH_GMAX,
        HATTERENE_GMAX,
        GRIMMSNARL_GMAX,
        ALCREMIE_GMAX,
        COPPERAJAH_GMAX,
        DURALUDON_GMAX,
        URSHIFU_SINGLE_STRIKE_GMAX,
        URSHIFU_RAPID_STRIKE_GMAX,
        TOXTRICITY_LOW_KEY_GMAX,
        GROWLITHE_HISUI,
        ARCANINE_HISUI,
        VOLTORB_HISUI,
        ELECTRODE_HISUI,
        TYPHLOSION_HISUI,
        QWILFISH_HISUI,
        SNEASEL_HISUI,
        SAMUROTT_HISUI,
        LILLIGANT_HISUI,
        ZORUA_HISUI,
        ZOROARK_HISUI,
        BRAVIARY_HISUI,
        SLIGGOO_HISUI,
        GOODRA_HISUI,
        AVALUGG_HISUI,
        DECIDUEYE_HISUI,
        DIALGA_ORIGIN,
        PALKIA_ORIGIN,
        BASCULIN_WHITE_STRIPED,
        BASCULEGION_FEMALE,
        ENAMORUS_THERIAN,
        TAUROS_PALDEA_COMBAT_BREED,
        TAUROS_PALDEA_BLAZE_BREED,
        TAUROS_PALDEA_AQUA_BREED,
        WOOPER_PALDEA,
        OINKOLOGNE_FEMALE,
        DUDUNSPARCE_THREE_SEGMENT,
        PALAFIN_HERO,
        MAUSHOLD_FAMILY_OF_THREE,
        TATSUGIRI_DROOPY,
        TATSUGIRI_STRETCHY,
        SQUAWKABILLY_BLUE_PLUMAGE,
        SQUAWKABILLY_YELLOW_PLUMAGE,
        SQUAWKABILLY_WHITE_PLUMAGE,
        GIMMIGHOUL_ROAMING,
        KORAIDON_LIMITED_BUILD,
        KORAIDON_SPRINTING_BUILD,
        KORAIDON_SWIMMING_BUILD,
        KORAIDON_GLIDING_BUILD,
        MIRAIDON_LOW_POWER_MODE,
        MIRAIDON_DRIVE_MODE,
        MIRAIDON_AQUATIC_MODE,
        MIRAIDON_GLIDE_MODE
      ]